+++ verto/dim3.py
@@ -40,8 +40,7 @@ class PcaEV(Pipeline):
         - transformed outputs are memory-friendly (np.float16),
         - create "feature_names_" output
     """
-    def __init__(self, required_ev=0.1, prefix="pcaev"):
-        self.prefix = prefix
+    def __init__(self, required_ev=0.1):
         self.required_ev = required_ev
         super().__init__(steps=[
             ('scl', StandardScaler(with_mean=True, with_std=True, copy=True)),
@@ -51,8 +50,8 @@ class PcaEV(Pipeline):
         ])
 
     def __str__(self):
-        return 'PcaEV(required_ev={}, steps={}, prefix={})'.format(
-            self.required_ev, self.steps, self.prefix)
+        return 'PcaEV(required_ev={}, steps={})'.format(
+            self.required_ev, self.steps)
 
     def __repr__(self):
         return self.__str__()
@@ -65,12 +64,12 @@ class PcaEV(Pipeline):
     def transform(self, X, y=None):
         return super().transform(X).astype(np.float16)
 
-    def fit(self, X, y=None):
-        super().fit(X)
-        self.feature_names_ = [
-            self.prefix + "_" + str(i) for i in
-            range(self.steps[1][1].n_components_)]
-        return self
+    # def fit(self, X, y=None):
+    #    super().fit(X)
+    #    self.feature_names_ = [
+    #        self.prefix + "_" + str(i) for i in
+    #        range(self.steps[1][1].n_components_)]
+    #    return self
 
 
 trans = PcaEV(required_ev=0.1)
@@ -83,5 +82,22 @@ meta = {
         "a required Explained Variance threshold"),
     'keywords': [
         'dimensionality reduction', 'principal component anlysis',
-        'StandardScaler', 'PCA', 'Explained Variance']
+        'StandardScaler', 'PCA', 'Explained Variance'],
+    'feature_names_prefix': 'pcaev'
 }
+
+
+"""Example
+
+from verto.dim3 import trans, meta
+from datasets.demo1 import X_train
+from seasalt import create_feature_names
+import pandas as pd
+import numpy as np
+
+trans.set_params(**{'required_ev': 0.8})
+X_new = trans.fit_transform(X_train).astype(np.float16)
+names = create_feature_names(meta['feature_names_prefix'], X_new.shape[1])
+df = pd.DataFrame(data=X_new, columns=names)
+df
+"""